pyproject.toml
[tool.poetry]
name = "skymarket"
version = "0.1.0"
description = "Course work 6"
authors = ["maximgrinin <maximgrinin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.6"
psycopg2 = "^2.9.5"
drf-spectacular = "^0.25.1"
django-cors-headers = "^3.13.0"
python-dotenv = "^0.21.1"
django-phonenumber-field = "^7.0.2"
djoser = "^2.1.0"
phonenumbers = "^8.13.5"
django-filter = "^22.1"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
